requirements.txt
PyQt5===5.15.10
requests===2.31.0
openai===1.13.3
pillow===10.2.0